Restore CodeQL warning fixes reverted by ICU 78 upstream import (re-apply #153)
Commit ba289fd3 ('Fixed CodeQL warnings (#153)') fixed 35 CodeQL findings across 23 files (tests, shipping common/ code, and tools). It was a direct commit, never recorded in icu-patches/, so the ICU 78 upstream source import (73bd9be6) silently reverted it. This re-applies those fixes onto the ICU 78 tree (22 files; wrtxml.cpp already carried the fix upstream). Optical-modernization conflicts (NULL->nullptr, C-casts->static_cast, UChar->char16_t) were resolved to the modern form plus the original logical fix.

+++ icu/icu4c/source/common/ucnvsel.cpp
@@ -725,7 +725,7 @@ static UEnumeration *selectForMask(const UConverterSelector* sel,
     int16_t k = 0;
     for (j = 0 ; j < columns; j++) {
       uint32_t v = mask[j];
-      for (i = 0 ; i < 32 && k < sel->encodingsCount; i++, k++) {
+      for (i = 0 ; i < 32 && k < static_cast<int16_t>(sel->encodingsCount); i++, k++) {
         if ((v & 1) != 0) {
           result->index[result->length++] = k;
         }
